requirements.txt
pandas>=2.0
numpy>=1.24
scikit-learn>=1.3
xgboost>=2.0
catboost>=1.2
lightgbm>=4.0
tensorflow>=2.15
joblib>=1.3
plotly>=5.18
dash>=2.14
dash-bootstrap-components>=1.5
transformers>=4.36
torch>=2.1
feedparser>=6.0
requests>=2.31
pyarrow>=14.0
scipy>=1.11
smartapi-python>=1.5
pyotp>=2.9
logzero>=1.7
websocket-client>=1.6
yfinance>=0.2
